requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
pony==0.7.6
python-dotenv==0.9.1
Werkzeug==0.14.1
